Created utils folder and moved logger into it
